Intermediate commit; troubleshooting react state update error.
