+++ frontend/src/pages/Risk/Risk.tsx
@@ -1,13 +1,13 @@
 import React, { useCallback, useState, useEffect } from 'react';
 import classes from './Risk.module.scss';
 import VulnerabilityCard from './VulnerabilityCard';
+import BarChartCardSmall from './BarChartCardSmall';
+import BarChartCardLarge from './BarChartCardLarge';
 import VulnerabilityPieChart from './VulnerabilityPieChart';
 import { useRiskStyles } from './style';
-import { getSingleColor, getSeverityColor, offsets, resultsPerPage } from './utils';
-import { ResponsiveBar } from '@nivo/bar';
+import { sevLabels, getSeverityColor, offsets, severities } from './utils';
 import { useAuthContext } from 'context';
-import { Paper, Chip } from '@material-ui/core';
-import { Pagination } from '@material-ui/lab';
+import { Paper } from '@material-ui/core';
 import { geoCentroid } from 'd3-geo';
 import {
   ComposableMap,
@@ -18,7 +18,6 @@ import {
   Annotation
 } from 'react-simple-maps';
 import { scaleLinear } from 'd3-scale';
-import { Link, useHistory } from 'react-router-dom';
 import { Vulnerability } from 'types';
 import { jsPDF } from 'jspdf';
 import html2canvas from 'html2canvas';
@@ -30,7 +29,7 @@ export interface Point {
   value: number;
 }
 
-interface Stats {
+export interface Stats {
   domains: {
     services: Point[];
     ports: Point[];
@@ -45,7 +44,7 @@ interface Stats {
   };
 }
 
-interface ApiResponse {
+export interface ApiResponse {
   result: Stats;
 }
 
@@ -53,7 +52,7 @@ interface VulnerabilityCount extends Vulnerability {
   count: number;
 }
 
-interface VulnSeverities {
+export interface VulnSeverities {
   label: string;
   sevList: string[];
   disable?: boolean;
@@ -66,21 +65,11 @@ let colorScale = scaleLinear<string>()
   .range(['#c7e8ff', '#135787']);
 
 const Risk: React.FC = (props) => {
-  const history = useHistory();
   const { currentOrganization, showAllOrganizations, showMaps, user, apiPost } =
     useAuthContext();
 
   const [stats, setStats] = useState<Stats | undefined>(undefined);
-  const [labels, setLabels] = useState([
-    '',
-    'Low',
-    'Medium',
-    'High',
-    'Critical'
-  ]);
-  const [domainsWithVulns, setDomainsWithVulns] = useState(0);
   const [isLoading, setIsLoading] = useState(false);
-  const [current, setCurrent] = useState(1);
   const cardClasses = useRiskStyles(props);
 
   const geoStateUrl = 'https://cdn.jsdelivr.net/npm/us-atlas@3/states-10m.json';
@@ -114,154 +103,6 @@ const Risk: React.FC = (props) => {
   useEffect(() => {
     fetchStats();
   }, [fetchStats]);
-
-  const VulnerabilityBarChart = ({
-    data,
-    xLabels,
-    type,
-    longXValues = false
-  }: {
-    data: Point[];
-    xLabels: string[];
-    type: string;
-    longXValues?: boolean;
-  }) => {
-    const keys = xLabels;
-    let dataVal: object[];
-    const pageStart = (current - 1) * resultsPerPage;
-    if (type === 'ports') {
-      dataVal = data.map((e) => ({ ...e, [xLabels[0]]: e.value })) as any;
-    } else {
-      // Separate count by severity
-      const domainToSevMap: any = {};
-      for (const point of data) {
-        const split = point.id.split('|');
-        const domain = split[0];
-        const severity = split[1];
-        if (labels.includes(severity)) {
-          if (!(domain in domainToSevMap)) domainToSevMap[domain] = {};
-          domainToSevMap[domain][severity] = point.value;
-        }
-      }
-      setDomainsWithVulns(Object.keys(domainToSevMap).length);
-      dataVal = Object.keys(domainToSevMap)
-        .map((key) => ({
-          label: key,
-          ...domainToSevMap[key]
-        }))
-        .sort((a, b) => {
-          let diff = 0;
-          for (const label of xLabels) {
-            diff += (label in b ? b[label] : 0) - (label in a ? a[label] : 0);
-          }
-          return diff;
-        })
-        .slice(pageStart, Math.min(pageStart + 30, domainsWithVulns))
-        .reverse();
-    }
-    // create the total vuln labels for each domain
-    const totalLabels = ({ bars, width }: any) => {
-      const fullWidth = width + 5;
-      return bars.map(
-        ({ data: { data, indexValue }, y, height, width }: any, i: number) => {
-          const total = Object.keys(data)
-            .filter((key) => key !== 'label')
-            .reduce((a, key) => a + data[key], 0);
-          if (i < dataVal.length) {
-            return (
-              <g
-                transform={`translate(${fullWidth}, ${y})`}
-                key={`${indexValue}-${i}`}
-              >
-                <text
-                  x={10}
-                  y={height / 2}
-                  textAnchor="middle"
-                  alignmentBaseline="central"
-                  // add any style to the label here
-                  style={{
-                    fill: 'rgb(51, 51, 51)',
-                    fontSize: 12
-                  }}
-                >
-                  {total}
-                </text>
-              </g>
-            );
-          }
-          return null;
-        }
-      );
-    };
-    return (
-      <ResponsiveBar
-        data={dataVal as any}
-        keys={keys}
-        layers={
-          type === 'ports'
-            ? ['grid', 'axes', 'bars', 'markers', 'legends']
-            : ['grid', 'axes', 'bars', totalLabels, 'markers', 'legends']
-        }
-        indexBy="label"
-        margin={{
-          top: longXValues ? 10 : 30,
-          right: 40,
-          bottom: longXValues ? 150 : 75,
-          left: longXValues ? 260 : 100
-        }}
-        theme={{
-          fontSize: 12,
-          axis: {
-            legend: {
-              text: {
-                fontWeight: 'bold'
-              }
-            }
-          }
-        }}
-        onClick={(event) => {
-          if (type === 'vulns') {
-            history.push(
-              `/inventory/vulnerabilities?domain=${event.data.label}&severity=${event.id}`
-            );
-          } else if (type === 'ports') {
-            history.push(
-              `/inventory?filters[0][field]=services.port&filters[0][values][0]=n_${event.data.label}_n&filters[0][type]=any`
-            );
-            window.location.reload();
-          }
-        }}
-        padding={0.5}
-        colors={type === 'ports' ? getSingleColor : (getSeverityColor as any)}
-        borderColor={{ from: 'color', modifiers: [['darker', 1.6]] }}
-        axisTop={null}
-        axisRight={null}
-        axisBottom={{
-          tickSize: 0,
-          tickPadding: 5,
-          tickRotation: 0,
-          legend: type === 'ports' ? 'Count' : '',
-          legendPosition: 'middle',
-          legendOffset: 40
-        }}
-        axisLeft={{
-          tickSize: 0,
-          tickPadding: 20,
-          tickRotation: 0,
-          legend: type === 'ports' ? 'Port' : '',
-          legendPosition: 'middle',
-          legendOffset: -65
-        }}
-        animate={true}
-        enableLabel={false}
-        motionDamping={15}
-        layout={'horizontal'}
-        enableGridX={true}
-        enableGridY={false}
-        {...({ motionStiffness: 90 } as any)}
-      />
-    );
-  };
 
   const MapCard = ({
     title,
@@ -366,15 +207,6 @@ const Risk: React.FC = (props) => {
     (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
   );
 
-  // Create severity object for Open Vulnerability chips
-  const severities: VulnSeverities[] = [
-    { label: 'All', sevList: ['', 'Low', 'Medium', 'High', 'Critical'] },
-    { label: 'Critical', sevList: ['Critical'] },
-    { label: 'High', sevList: ['High'] },
-    { label: 'Medium', sevList: ['Medium'] },
-    { label: 'Low', sevList: ['Low'] }
-  ];
-
   if (stats) {
     for (const sev of severities) {
       sev.disable = !stats.domains.numVulnerabilities.some((i) =>
@@ -462,7 +294,7 @@ const Risk: React.FC = (props) => {
                       <h2>Most common ports</h2>
                     </div>
                     <div className={cardClasses.chartSmall}>
-                      <VulnerabilityBarChart
+                      <BarChartCardSmall
                         data={stats.domains.ports.slice(0, 5).reverse()}
                         type={'ports'}
                         xLabels={['Port']}
@@ -485,81 +317,11 @@ const Risk: React.FC = (props) => {
               <Paper elevation={0} className={cardClasses.cardRoot}>
                 <div>
                   {stats.domains.numVulnerabilities.length > 0 && (
-                    <div className={cardClasses.cardBig}>
-                      <div className={cardClasses.seeAll}>
-                        <h4>
-                          <Link to="/inventory/vulnerabilities">See All</Link>
-                        </h4>
-                      </div>
-                      <div className={cardClasses.header}>
-                        <h2>Open Vulnerabilities by Domain</h2>
-                      </div>
-                      <div className={cardClasses.chartLarge}>
-                        {stats.domains.numVulnerabilities.length === 0 ? (
-                          <h3>No open vulnerabilities</h3>
-                        ) : (
-                          <>
-                            <p className={cardClasses.note}>
-                              *Top 50 domains with open vulnerabilities
-                            </p>
-                            <div className={cardClasses.chipWrapper}>
-                              {severities.map(
-                                (sevFilter: VulnSeverities, i: number) => (
-                                  <Chip
-                                    key={i}
-                                    className={cardClasses.chip}
-                                    disabled={sevFilter.disable}
-                                    label={sevFilter.label}
-                                    onClick={() => {
-                                      setLabels(sevFilter.sevList);
-                                      setCurrent(1);
-                                    }}
-                                  ></Chip>
-                                )
-                              )}
-                            </div>
-                            <div className={cardClasses.chartHeader}>
-                              <h5>Domain&emsp; Breakdown</h5>
-                              <h5
-                                style={{ textAlign: 'right', paddingLeft: 0 }}
-                              >
-                                Total
-                              </h5>
-                            </div>
-                            <VulnerabilityBarChart
-                              data={stats.domains.numVulnerabilities}
-                              xLabels={labels}
-                              type={'vulns'}
-                              longXValues={true}
-                            />
-                          </>
-                        )}
-                      </div>
-                      <div className={cardClasses.footer}>
-                        <span>
-                          <strong>
-                            {(domainsWithVulns === 0
-                              ? 0
-                              : (current - 1) * resultsPerPage + 1
-                            ).toLocaleString()}{' '}
-                            -{' '}
-                            {Math.min(
-                              (current - 1) * resultsPerPage + resultsPerPage,
-                              domainsWithVulns
-                            ).toLocaleString()}
-                          </strong>{' '}
-                          of{' '}
-                          <strong>{domainsWithVulns.toLocaleString()}</strong>
-                        </span>
-                        <Pagination
-                          count={Math.ceil(domainsWithVulns / resultsPerPage)}
-                          page={current}
-                          onChange={(_, page) => setCurrent(page)}
-                          color="primary"
-                          size="small"
-                        />
-                      </div>
-                    </div>
+                    <BarChartCardLarge
+                      data={stats.domains.numVulnerabilities}
+                      type={'vulns'}
+                      xLabels={sevLabels}
+                    />
                   )}
                 </div>
               </Paper>
